SAK-28947; fix Longsight code to build; add BB samples
git-svn-id: https://source.sakaiproject.org/svn/common/trunk@316276 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ common/import-parsers/blackboard_6/impl/src/java/org/sakaiproject/importer/impl/translators/Bb6AssessmentTranslator.java
@@ -262,9 +262,9 @@ public class Bb6AssessmentTranslator implements IMSResourceTranslator{
 			// Drop comment garbage
 			q.setQuestionText(stripHTMLComments(questionTextString + questionResourceEmbed));
 			try {
-				float pointValue = Float.parseFloat(XPathHelper.getNodeValue("./itemmetadata/qmd_absolutescore_max", questionNode));
+				double pointValue = Double.parseDouble(XPathHelper.getNodeValue("./itemmetadata/qmd_absolutescore_max", questionNode));
 				if (pointValue < 0) {
-					pointValue = Float.parseFloat(XPathHelper.getNodeValue("./resprocessing/outcomes/decvar/@maxvalue", questionNode));
+					pointValue = Double.parseDouble(XPathHelper.getNodeValue("./resprocessing/outcomes/decvar/@maxvalue", questionNode));
 				}
 				q.setPointValue(pointValue);
 			} catch (NumberFormatException e) {
